2_lead_enrichment_marketing/requirements.txt
langchain
langchain-openai
langchain-community
beautifulsoup4
pandas
googlesearch-python
python-dotenv
requests
streamlit
